+++ arknights_mower/__main__.py
@@ -6,7 +6,7 @@ from arknights_mower.solvers.reclamation_algorithm import ReclamationAlgorithm
 from arknights_mower.solvers.secret_front import SecretFront
 from arknights_mower.utils import config, path, rapidocr
 from arknights_mower.utils.csleep import MowerExit
-from arknights_mower.utils.datetime import format_time, get_server_time
+from arknights_mower.utils.datetime_utils import format_time, get_server_time
 from arknights_mower.utils.depot import 创建csv, 创建json
 from arknights_mower.utils.device.adb_client.session import Session
 from arknights_mower.utils.device.scrcpy import Scrcpy
